fixed view/data type issue with temperature

+++ src/main/java/hmmelton/com/stormy/CurrentWeather.java
@@ -87,7 +87,7 @@ public class CurrentWeather {
         mTime = time;
     }
 
-    public double getTemperature() {
+    public int getTemperature() {
         return (int) Math.round(mTemperature);
     }
